+++ CHANGELOG.md
@@ -4,11 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2021-03-27
+## [Unreleased] - 2021-03-29
 
 ### Added
+* New methods added to the Permutation class.
+    * The cycle method creates a permutation cycle.
+    * Multiple variations of a scramble(int[] indexes) method that 
+      randomizes the elements indicated by the array of indexes. The
+      variations of this method correspond to the different random number
+      generators supported by the library.
 
 ### Changed
+* Minor code improvements in Permutation class (e.g., refactoring to 
+  reduce duplicated code, and some minor code optimizations).
 
 ### Deprecated